chore: fix PR comments
