+++ components/HomepageSection/Activity.jsx
@@ -1,7 +1,4 @@
-import {
-  getA2ATransactions,
-  getUniqueOperatorCount,
-} from 'common-util/api/dune';
+import { getA2ATransactions, getFeeFlowMetrics } from 'common-util/api/dune';
 import {
   get7DaysAvgActivity,
   getTotalTransactionsCount,
@@ -10,203 +7,371 @@ import {
 import {
   DUNE_A2A_TRANSACTIONS_QUERY_URL,
   DUNE_AGENTS_QUERY_URL,
+  DUNE_MMV2_URL,
   FLIPSIDE_DAAS_QUERY_URL,
   FLIPSIDE_URL,
 } from 'common-util/constants';
-import SectionWrapper from 'components/Layout/SectionWrapper';
+import SectionHeading from 'components/SectionHeading';
 import { Card } from 'components/ui/card';
 import { Popover } from 'components/ui/popover';
 import { ExternalLink } from 'components/ui/typography';
-import chains from 'data/chains.json';
 import { usePersistentSWR } from 'hooks';
 import Image from 'next/image';
 import Link from 'next/link';
 import { useMemo } from 'react';
-import SectionHeading from '../SectionHeading';
 
-const BLOCKCHAIN_COUNT = chains.length;
+const imgPath = '/images/homepage/activity/';
 
 const fetchMetrics = async () => {
-  const [transactions, agents, unitsCount, dailyActiveAgents, a2aTransactions] =
+  const [transactions, agents, dailyActiveAgents, a2aTransactions, feeMetrics] =
     await Promise.allSettled([
       getTotalTransactionsCount(),
-      getUniqueOperatorCount(),
       getTotalUnitsCount(),
       get7DaysAvgActivity(),
       getA2ATransactions(),
+      getFeeFlowMetrics(),
     ]);
 
   return {
     transactions:
       transactions.status === 'fulfilled' ? transactions.value : null,
     agents: agents.status === 'fulfilled' ? agents.value : null,
-    agentsTypes: unitsCount.status === 'fulfilled' ? unitsCount.value : null,
     dailyActiveAgents:
       dailyActiveAgents.status === 'fulfilled' ? dailyActiveAgents.value : null,
     a2aTransactions:
       a2aTransactions.status === 'fulfilled' ? a2aTransactions.value : null,
+    feeMetrics: feeMetrics.status == 'fulfilled' ? feeMetrics.value : null,
   };
 };
 
-export const Activity = () => {
-  const { data: metrics } = usePersistentSWR('activityMetrics', fetchMetrics);
+const agents = ['predict', 'babydegen', 'mech', 'agentsfun'];
 
-  const data = useMemo(
-    () => [
-      {
-        id: 'transactions',
-        subText: 'transactions made by Olas agents',
-        value: metrics?.transactions?.toLocaleString(),
-        source: FLIPSIDE_URL,
-        isExternal: true,
-        btmContent: (
-          <div className="text-sm mt-4">
-            <p className="text-slate-500 font-medium">of which</p>
-            {metrics?.a2aTransactions ? (
-              <ExternalLink
-                className="text-2xl font-bold text-purple-600"
-                href={DUNE_A2A_TRANSACTIONS_QUERY_URL}
-              >
-                {metrics.a2aTransactions.toLocaleString()}
-              </ExternalLink>
-            ) : (
-              <span className="text-purple-600 text-2xl">--</span>
-            )}
-            <p>agent-to-agent transactions</p>
-          </div>
-        ),
-      },
-      {
-        id: 'agentsTypes',
-        subText: 'types of agents registered by Builders',
-        value: metrics?.agentsTypes,
-        source: `${FLIPSIDE_URL}?tabIndex=5`,
-        isExternal: true,
-      },
-      {
-        id: 'agents',
-        subText: 'agents deployed by Operators',
-        value: metrics?.agents,
-        source: DUNE_AGENTS_QUERY_URL,
-        isExternal: true,
-      },
-      {
-        id: 'blockchains',
-        subText: 'blockchains Olas is deployed across',
-        value: BLOCKCHAIN_COUNT,
-        source: '/#chains',
-        isExternal: false,
-        btmContent: (
-          <div className="mt-16 flex flex-row justify-between w-full">
-            <Image
-              key="Chains"
-              src={`/images/homepage/chain-logos.png`}
-              alt="Activity"
-              width={340}
-              height={25}
-              className="w-full"
-            />
-          </div>
-        ),
-      },
-    ],
-    [metrics],
-  );
+const OlasIsBurnedArrow = ({ pointsDown = false, className }) => (
+  <div className={`flex flex-row md:mt-4 gap-2 ${className}`}>
+    <p className="w-[76px] md:ml-[58px] text-sm mt-9 md:mt-[46px] mb-auto max-sm:text-slate-500">
+      OLAS is burned
+    </p>
+    <Image
+      src={`${imgPath}${pointsDown ? 'mobile-arrow4.png' : 'arrow4.png'}`}
+      alt="arrow"
+      width={50}
+      height={124}
+      className="h-[124px]"
+    />
+    <div className="mb-auto mt-[47px] font-semibold text-sm md:mt-[58px] text-black z-20 content-center flex flex-row gap-2">
+      <p>OFF</p>
+      <Popover
+        align="center"
+        side="right"
+        contentClassName="w-[382px] text-left font-normal translate-x-2"
+      >
+        Fees collected can be turned on or off by the Governors of the Olas
+        Protocol. Currently, fees are turned off to encourage early adoption and
+        growth of the marketplace.
+        <ExternalLink className="mt-2 cursor-pointer">
+          More about Mech Marketplace fees in AIP-5
+        </ExternalLink>
+      </Popover>
+    </div>
+  </div>
+);
+
+const ActivityCard = ({
+  icon,
+  iconWidth = 40,
+  iconHeight = 40,
+  text,
+  primary: { text: primaryText, link: primaryLink, value: primaryValue },
+  secondary = {},
+}) => {
+  const {
+    text: secondaryText,
+    link: secondaryLink,
+    value: secondaryValue,
+  } = secondary;
 
   return (
-    <SectionWrapper
-      customClasses="text-center border-b"
-      id="generating-decentralized-autonomous-ai-agent-economies"
-    >
-      <div className="text-7xl lg:text-9xl max-w-[900px] mx-auto pt-16 px-4">
+    <Card className="flex flex-col py-4 px-6 gap-4 h-fit w-full md:w-[300px] activity-card-opaque">
+      <div className="flex flex-row place-items-center gap-3">
         <Image
-          alt="Placeholder"
-          src="/images/generating-agents.png"
-          width="270"
-          height="300"
-          className="mx-auto mb-16"
+          src={`${imgPath}${icon}`}
+          alt={text}
+          width={iconWidth}
+          height={iconHeight}
         />
-        <SectionHeading
-          size="lg:text-5xl"
-          color="text-gray-900"
-          weight="font-bold"
-        >
-          Generating Decentralized Autonomous AI Agent Economies
-        </SectionHeading>
-        <p className="text-lg md:text-xl text-slate-600 mb-8 mx-auto">
-          Olas incentivizes the creation of open-source autonomous AI agents,
-          enabling users to own agents for their benefit and businesses to
-          monetize agent-to-agent services. As a decentralized AI
-          infrastructure,{' '}
-          <Link
-            href="https://staking.olas.network/"
-            className="text-purple-600"
-          >
-            Olas Staking
-          </Link>{' '}
-          rewards agents for contributions and coordinates expanding agent
-          economies.
-        </p>
-        <Card className="flex flex-col items-center gap-6 p-8 mb-16 border border-purple-200 rounded-full text-xl w-fit mx-auto rounded-2xl bg-gradient-to-t from-[#F1DBFF] to-[#FDFAFF]">
-          <span>
-            🤖 The first autonomous AI agents are{' '}
-            <span className="font-medium">active</span>
-          </span>
-          {metrics?.dailyActiveAgents ? (
-            <ExternalLink
-              className="font-extrabold text-6xl"
-              href={FLIPSIDE_DAAS_QUERY_URL}
-              hideArrow
-            >
-              {metrics.dailyActiveAgents}
-              <span className="text-4xl">↗</span>
-            </ExternalLink>
-          ) : (
-            <span className="text-purple-600 text-6xl">--</span>
-          )}
-          <div className="flex self-center gap-2">
-            Daily Active Agents (DAAs)
-            <Popover>7-day average Daily Active Agents</Popover>
-          </div>
-        </Card>
+        {text}
       </div>
-      <div className="w-full bg-gradient-to-t from-gray-100 to-gray-100/0 relative max-sm:h-[850px]">
-        <div className="absolute z-0 bg-[url('/images/homepage/metric-bg.png')] opacity-40 h-full w-full"></div>
-        <div className="py-12">
-          <p className="text-xl md:text-2xl text-slate-700 mb-8 font-semibold mx-auto">
-            Olas agent economies show a growing lifetime traction
-          </p>
-          <div className="flex flex-col max-sm:flex-row flex-wrap gap-6 mx-auto max-w-[800px] max-h-[350px]">
-            {data.map((item) => {
-              const getValue = () => {
-                if (!item.value) return '--';
-                if (item.isExternal) {
-                  return (
-                    <ExternalLink href={item.source} hideArrow>
-                      {item.value}
-                      <span className="text-2xl">↗</span>
-                    </ExternalLink>
-                  );
-                }
-                return <Link href={item.source}>{item.value}</Link>;
-              };
+      <div className="flex flex-row gap-2 place-items-center">
+        <ExternalLink href={primaryLink}>
+          <div className="text-purple-700 text-2xl font-semibold">
+            {primaryValue}
+          </div>
+        </ExternalLink>
+        {primaryText}
+      </div>
+      {secondaryValue && (
+        <div className="flex flex-row gap-2 place-items-center">
+          <ExternalLink href={secondaryLink}>
+            <div className="text-purple-700 text-xl font-semibold">
+              {secondaryValue}
+            </div>
+          </ExternalLink>
+          {secondaryText}
+        </div>
+      )}
+    </Card>
+  );
+};
 
-              return (
-                <Card
-                  key={item.id}
-                  className="z-10 max-w-[388px] max-sm:mx-auto max-sm:w-[400px] border-white bg-gradient-to-t from-slate-100/40 to-white/50 to-40% shadow-lg text-start px-6 py-4 rounded-xl"
-                >
-                  <span className="block text-4xl font-bold text-purple-600">
-                    {getValue()}
-                  </span>
-                  <span>{item.subText}</span>
-                  {item.btmContent}
-                </Card>
-              );
-            })}
+export const Activity = () => {
+  const { data: metrics } = usePersistentSWR('tokenMetrics', fetchMetrics);
+
+  const processedMetrics = useMemo(() => {
+    if (!metrics) return null;
+
+    return {
+      transactions: metrics.transactions?.toLocaleString() || '--',
+      agents: metrics.agents?.toLocaleString() || '--',
+      dailyActiveAgents: metrics.dailyActiveAgents?.toLocaleString() || '--',
+      a2aTransactions: metrics.a2aTransactions?.toLocaleString() || '--',
+      feesCollected: metrics.feeMetrics?.totalFees?.toFixed(2) || '--',
+      olasBurned: metrics.feeMetrics?.olasBurned?.toFixed(2) || '--',
+    };
+  }, [metrics]);
+
+  return (
+    <div className="max-w-4xl mx-auto">
+      <SectionHeading
+        size="max-sm:text-5xl"
+        color="text-gray-900"
+        weight="font-bold"
+        other="mb-12 max-w-3xl text-center mx-auto max-lg:mx-4"
+      >
+        OLAS: Powers AI Agent Economies
+      </SectionHeading>
+      <p className="text-lg text-slate-600 mb-20 max-lg:mx-4">
+        The OLAS token is bootstrapping a flywheel birthing larger and larger
+        agent economies: Each Pearl user stakes OLAS to access their
+        agents&apos; benefits. To provide utility to their users Pearl agents
+        use the marketplace. The marketplace charges fees to agents. Fees are
+        used to burn OLAS.
+      </p>
+
+      <div className="flex flex-col max-w-4xl mx-auto text-slate-500 place-items-center hidden md:flex">
+        <div className="flex flex-row">
+          <p className="w-[124px] text-sm pt-24">
+            Attracts more builders and users
+          </p>
+          <Image
+            src={`${imgPath}arrow5.png`}
+            alt="arrow"
+            width={124}
+            height={176}
+            className="mt-12 ml-3 md:mr-4 w-[124px] h-[176px]"
+          />
+          <ActivityCard
+            icon="users.png"
+            text="Users"
+            primary={{
+              value: processedMetrics?.agents,
+              text: 'agents deployed',
+              link: DUNE_AGENTS_QUERY_URL,
+            }}
+          />
+          <Image
+            src={`${imgPath}arrow.png`}
+            alt="arrow"
+            width={150}
+            height={142}
+            className="mt-14 mr-2 w-[150px] h-[142px]"
+          />
+          <p className="w-[124px] text-sm pt-24">
+            Stake OLAS to use Pearl agents
+          </p>
+        </div>
+        <div className="flex flex-row place-items-center w-full justify-between">
+          <div className="flex flex-col">
+            <ActivityCard
+              icon="olas-burnt.png"
+              primary={{
+                value: `$${Number(processedMetrics?.olasBurned).toLocaleString()}`,
+                text: 'OLAS burned',
+                link: DUNE_MMV2_URL,
+              }}
+            />
+            <OlasIsBurnedArrow />
+          </div>
+          <div className="flex flex-col place-items-center z-10">
+            <div className="flex flex-row w-[124px] flex-wrap mb-2 px-auto">
+              {agents.map((item) => (
+                <Image
+                  key={item}
+                  src={`/images/homepage/activity/${item}.png`}
+                  alt={item}
+                  width={62}
+                  height={62}
+                  className="hover:-translate-y-1 duration-150"
+                />
+              ))}
+            </div>
+            <Link
+              href="/agent-economies"
+              className="text-purple-700 hover:text-purple-900"
+            >
+              Agent economies
+            </Link>
+          </div>
+          <div className="flex flex-col">
+            <ActivityCard
+              icon="daas.png"
+              iconWidth={252}
+              iconHeight={56}
+              primary={{
+                value: processedMetrics?.dailyActiveAgents,
+                text: (
+                  <>
+                    DAAs <Popover>7-day average Daily Active Agents</Popover>
+                  </>
+                ),
+                link: FLIPSIDE_DAAS_QUERY_URL,
+              }}
+            />
+            <div className="flex flex-row">
+              <Image
+                src={`${imgPath}arrow2.png`}
+                alt="arrow2"
+                width={28}
+                height={124}
+                className="h-[124px] ml-auto"
+              />
+              <p className="w-[76px] text-sm mr-[68px] ml-2 my-auto">
+                Agents are active
+              </p>
+            </div>
           </div>
         </div>
+        <div className="flex flex-row place-items-center">
+          <ActivityCard
+            icon="agent-to-agent.png"
+            iconWidth={104}
+            iconHeight={36}
+            primary={{
+              value: processedMetrics?.a2aTransactions,
+              text: 'A2A txns',
+              link: DUNE_A2A_TRANSACTIONS_QUERY_URL,
+            }}
+            secondary={{
+              value: `$${Number(processedMetrics?.feesCollected).toLocaleString()}`,
+              text: 'fees collected',
+              link: DUNE_MMV2_URL,
+            }}
+          />
+          <div>
+            <Image
+              src={`${imgPath}arrow3.png`}
+              alt="arrow2"
+              width={257}
+              height={10}
+              className="ml-4"
+            />
+            <p>AI Agent Bazaar is used</p>
+          </div>
+          <ActivityCard
+            icon="txns.png"
+            primary={{
+              value: processedMetrics?.transactions,
+              text: 'txns',
+              link: FLIPSIDE_URL,
+            }}
+          />
+        </div>
       </div>
-    </SectionWrapper>
+
+      <div className="flex flex-col md:hidden w-[90%] mx-auto">
+        <ActivityCard
+          icon="users.png"
+          text="Users"
+          primary={{
+            value: processedMetrics?.agents,
+            text: 'agents deployed',
+            link: DUNE_AGENTS_QUERY_URL,
+          }}
+        />
+        <Image
+          src={`${imgPath}mobile-arrow.png`}
+          alt="arrow"
+          width={240}
+          height={120}
+          className="mx-auto mb-2"
+        />
+        <ActivityCard
+          icon="daas.png"
+          iconWidth={252}
+          iconHeight={56}
+          primary={{
+            value: processedMetrics?.dailyActiveAgents,
+            text: (
+              <>
+                DAAs <Popover>7-day average Daily Active Agents</Popover>
+              </>
+            ),
+            link: FLIPSIDE_DAAS_QUERY_URL,
+          }}
+        />
+        <Image
+          src={`${imgPath}mobile-arrow2.png`}
+          alt="arrow"
+          width={132}
+          height={120}
+          className="mx-auto mb-2"
+        />
+        <ActivityCard
+          icon="txns.png"
+          primary={{
+            value: processedMetrics?.transactions,
+            text: 'txns',
+            link: FLIPSIDE_URL,
+          }}
+        />
+        <Image
+          src={`${imgPath}mobile-arrow3.png`}
+          alt="arrow"
+          width={180}
+          height={120}
+          className="mx-auto mb-2"
+        />
+        <ActivityCard
+          icon="agent-to-agent.png"
+          iconWidth={104}
+          iconHeight={36}
+          primary={{
+            value: processedMetrics?.a2aTransactions,
+            text: 'A2A txns',
+            link: DUNE_A2A_TRANSACTIONS_QUERY_URL,
+          }}
+          secondary={{
+            value: `$${Number(processedMetrics?.feesCollected).toLocaleString()}`,
+            text: 'fees collected',
+            link: DUNE_MMV2_URL,
+          }}
+        />
+        <OlasIsBurnedArrow pointsDown className="mx-auto mb-2" />
+        <ActivityCard
+          icon="olas-burnt.png"
+          primary={{
+            value: `$${Number(processedMetrics?.olasBurned).toLocaleString()}`,
+            text: 'OLAS burned',
+            link: DUNE_MMV2_URL,
+          }}
+        />
+        <Image
+          src={`${imgPath}mobile-arrow5.png`}
+          alt="arrow"
+          width={287}
+          height={202}
+          className="ml-[30px]"
+        />
+      </div>
+    </div>
   );
 };
