+++ styles/globals.css
@@ -84,7 +84,7 @@
   }
 }
 
-.homepage-hero-pg {
+.homepage-hero-bg {
   object-fit: contain;
   width: 100vw;
   top: 0;
